fix site->site_name for login
