+++ tabcmd/commands/datasources_and_workbooks/export_command.py
@@ -2,7 +2,6 @@ import tableauserverclient as TSC
 
 from tabcmd.commands.auth.session import Session
 from tabcmd.execution.logger_config import log
-from tabcmd.parsers.export_parser import ExportParser
 from .datasources_and_workbooks_command import DatasourcesAndWorkbooks
