+++ tests/test_login_parser.py
@@ -29,7 +29,7 @@ class LoginParserTest(unittest.TestCase):
         assert args.username == "me", args
         assert args.password == "the-password", args
         assert args.server == "the-server", args
-        assert args.site == "the-site"
+        assert args.site_name == "the-site"
 
     def test_parsing_connection_settings(self):
         mock_args = [commandname, "--no-certcheck", "--proxy", "the-proxy", "--timeout", "10"]
@@ -84,4 +84,21 @@ class LoginParserTest(unittest.TestCase):
         mock_args = [commandname, "--version", "the-pwd", "--password_file", "a-file-name"]
         with self.assertRaises(SystemExit):
             # this runs an action and exits
+            args = self.parser_under_test.parse_args(mock_args)
+
+    def test_login_parser_test_username_password(self):
+        mock_args = [commandname, "--username", "helloworld", "--password", "pw"]
+        args = self.parser_under_test.parse_args(mock_args)
+        assert args.username == "helloworld", args
+        assert args.password == "pw", args
+
+    def test_login_parser_test_token(self):
+        mock_args = [commandname, "--token", "token", "--token-name", "tn"]
+        args = self.parser_under_test.parse_args(mock_args)
+        assert args.token_name == "tn", args
+        assert args.token == "token", args
+
+    def test_login_token_and_username(self):
+        mock_args = [commandname, "--token-name", "to", "--username", "u"]
+        with self.assertRaises(SystemExit):
             args = self.parser_under_test.parse_args(mock_args)
